perf: cut encode wall time ~2.2x by parallelising parsing and removing redundant work
Encode of d2.grib2 --max-zoom 5 drops from ~480ms to ~215ms total
(tile/encode phase: 340ms -> ~28ms).

- parser: scan message offsets in pure Go (MessageRanges), then decode
  headers+values across GOMAXPROCS workers from in-memory bytes; collapses
  the old two-phase header-scan + data-stream pipeline into one cgo pass
- parser: switch values to float32 + codes_get_float_array (2x faster
  cgo transfer than codes_get_double_array on this input)
- parser: fix distinctLatitudes/Longitudes cache race - 24 workers were
  all missing the cold cache and redundantly calling eccodes; single-flight
  via a ready channel saved ~80ms on its own
- encoder: SkipInternalWorkers + DirectWorker run quantize+codec inline on
  caller goroutines, removing two channel hops and the serializer goroutine
- codec: SharedSampler shares the bandit state across workers so the
  triple-encode sampling phase converges once globally instead of once
  per (worker, variable)
- cmd/wmtiles: single-pass parseAllMessages + tileFromParsed, both ported
  through encode and extend

+++ tiler/tiler_test.go
@@ -18,10 +18,10 @@ func makeGrid() *parser.GRIBFile {
 	for j := range nx {
 		lons[j] = -10 + float64(j)
 	}
-	data := make([]float64, nx*ny)
+	data := make([]float32, nx*ny)
 	for r := range ny {
 		for c := range nx {
-			data[r*nx+c] = lats[r] + 0.1*lons[c]
+			data[r*nx+c] = float32(lats[r] + 0.1*lons[c])
 		}
 	}
 	return &parser.GRIBFile{
@@ -91,7 +91,7 @@ func TestNorthDownGridSampler(t *testing.T) {
 	}
 	g.Header.DistinctLatitudes = rev
 	rowBytes := g.Header.Nx
-	revData := make([]float64, len(g.DataValues))
+	revData := make([]float32, len(g.DataValues))
 	for r := range g.Header.Ny {
 		copy(revData[r*rowBytes:(r+1)*rowBytes],
 			g.DataValues[(g.Header.Ny-1-r)*rowBytes:(g.Header.Ny-r)*rowBytes])
